embed excel interop types (#14199)
* init

* Update DynamoCore.net6.sln

* Update DSOffice.csproj

* Update DSOffice.csproj

* Update DSOffice.csproj

* Update Dynamo.All.Net6.sln

* Revert "Update Dynamo.All.Net6.sln"

This reverts commit b7a919accfe22bbb8b0c85b36cec40d462f1ae93.

* Update DynamoCore.net6.sln

* update

* update

* Update DSOffice.csproj

* update

* Update FFITest.cs

* update

* Update Dynamo.All.Net6.sln

* remove net48 dsoffice

* update

* update

---------
