+++ src/Config/CS_SDK.props
@@ -52,9 +52,11 @@
   </PropertyGroup>
   <PropertyGroup Condition="$(Platform.Contains('Windows'))" >
     <RuntimeIdentifier>win10-x64</RuntimeIdentifier>
+    <DefineConstants>$(DefineConstants);_WINDOWS</DefineConstants>
   </PropertyGroup>
   <PropertyGroup Condition="$(Platform.Contains('Linux'))" >
     <RuntimeIdentifier>linux-x64</RuntimeIdentifier>
+    <DefineConstants>$(DefineConstants);_LINUX</DefineConstants>
   </PropertyGroup>
   <PropertyGroup Condition="'$(UILib)' == 'true' AND $(Platform.Contains('NET60'))">
     <TargetFramework>net6.0-windows</TargetFramework>
