+++ src/DynamoCore.sln
@@ -97,8 +97,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{5B9B5B6B
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestServices", "..\test\Libraries\TestServices\TestServices.csproj", "{6CD0F0CF-8199-49F9-B0EA-0B9598B44419}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DSOffice", "Libraries\DSOffice\DSOffice.csproj", "{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DynamoCoreTests", "..\test\DynamoCoreTests\DynamoCoreTests.csproj", "{472084ED-1067-4B2C-8737-3839A6143EB2}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FFITarget", "..\test\Engine\FFITarget\FFITarget.csproj", "{C70FE632-5500-4C57-B3D6-9B5574137551}"
@@ -120,6 +118,10 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GeometryColorTests", "..\test\Libraries\GeometryColorTests\GeometryColorTests.csproj", "{F8D2E3E5-4505-4463-9367-2EB2629D7DCD}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnitsNodeModels", "Libraries\UnitsNodeModels\UnitsNodeModels.csproj", "{61169E56-DED9-464F-B800-AE31AEC1F750}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DSOffice", "Libraries\DSOffice\DSOffice.csproj", "{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DSOfficeUtilities", "Libraries\DSOfficeUtilities\DSOfficeUtilities.csproj", "{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -369,14 +371,6 @@ Global
 		{6CD0F0CF-8199-49F9-B0EA-0B9598B44419}.Release|NET60_Linux.Build.0 = Release|NET60_Linux
 		{6CD0F0CF-8199-49F9-B0EA-0B9598B44419}.Release|NET60_Windows.ActiveCfg = Release|NET60_Windows
 		{6CD0F0CF-8199-49F9-B0EA-0B9598B44419}.Release|NET60_Windows.Build.0 = Release|NET60_Windows
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Debug|NET60_Linux.ActiveCfg = Debug|NET60_Linux
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Debug|NET60_Linux.Build.0 = Debug|NET60_Linux
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Debug|NET60_Windows.ActiveCfg = Debug|NET60_Windows
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Debug|NET60_Windows.Build.0 = Debug|NET60_Windows
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Release|NET60_Linux.ActiveCfg = Release|NET60_Linux
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Release|NET60_Linux.Build.0 = Release|NET60_Linux
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Release|NET60_Windows.ActiveCfg = Release|NET60_Windows
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Release|NET60_Windows.Build.0 = Release|NET60_Windows
 		{472084ED-1067-4B2C-8737-3839A6143EB2}.Debug|NET60_Linux.ActiveCfg = Debug|NET60_Linux
 		{472084ED-1067-4B2C-8737-3839A6143EB2}.Debug|NET60_Linux.Build.0 = Debug|NET60_Linux
 		{472084ED-1067-4B2C-8737-3839A6143EB2}.Debug|NET60_Windows.ActiveCfg = Debug|NET60_Windows
@@ -449,6 +443,20 @@ Global
 		{61169E56-DED9-464F-B800-AE31AEC1F750}.Release|NET60_Linux.Build.0 = Release|NET60_Linux
 		{61169E56-DED9-464F-B800-AE31AEC1F750}.Release|NET60_Windows.ActiveCfg = Release|NET60_Windows
 		{61169E56-DED9-464F-B800-AE31AEC1F750}.Release|NET60_Windows.Build.0 = Release|NET60_Windows
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}.Debug|NET60_Linux.ActiveCfg = Debug|NET60_Linux
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}.Debug|NET60_Linux.Build.0 = Debug|NET60_Linux
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}.Debug|NET60_Windows.ActiveCfg = Debug|NET60_Windows
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}.Debug|NET60_Windows.Build.0 = Debug|NET60_Windows
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}.Release|NET60_Linux.ActiveCfg = Release|NET60_Linux
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}.Release|NET60_Linux.Build.0 = Release|NET60_Linux
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}.Release|NET60_Windows.ActiveCfg = Release|NET60_Windows
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1}.Release|NET60_Windows.Build.0 = Release|NET60_Windows
+		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Debug|NET60_Linux.ActiveCfg = Debug|NET60_Linux
+		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Debug|NET60_Windows.ActiveCfg = Debug|NET60_Windows
+		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Debug|NET60_Windows.Build.0 = Debug|NET60_Windows
+		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Release|NET60_Linux.ActiveCfg = Release|NET60_Linux
+		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Release|NET60_Windows.ActiveCfg = Release|NET60_Windows
+		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7}.Release|NET60_Windows.Build.0 = Release|NET60_Windows
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -478,7 +486,6 @@ Global
 		{31183026-DE70-49CB-BC7C-0DFD0A088F62} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{06B9E5B0-7C50-4351-9D88-E159DC25755F} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{6CD0F0CF-8199-49F9-B0EA-0B9598B44419} = {5B9B5B6B-0BA7-4606-B8E5-70C958346D57}
-		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{472084ED-1067-4B2C-8737-3839A6143EB2} = {5B9B5B6B-0BA7-4606-B8E5-70C958346D57}
 		{C70FE632-5500-4C57-B3D6-9B5574137551} = {9739B26B-F223-458C-88AA-88D11E025EA0}
 		{2235F7CA-AB25-4177-92BC-4B574D2D5DF6} = {9739B26B-F223-458C-88AA-88D11E025EA0}
@@ -490,6 +497,8 @@ Global
 		{AB735028-22B0-4083-9CC5-805C582818EA} = {9CE88661-B9BD-4587-962A-AE0648B34EDA}
 		{F8D2E3E5-4505-4463-9367-2EB2629D7DCD} = {9CE88661-B9BD-4587-962A-AE0648B34EDA}
 		{61169E56-DED9-464F-B800-AE31AEC1F750} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
+		{6CC1E399-B5A5-4FA2-A98C-5D8B2BA280A1} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
+		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {89CB19C6-BF0A-4E6A-BFDA-79D143EAB59D}
