+++ src/Libraries/DSOfficeUtilities/CompatibilityMarshal.cs
