+++ src/Libraries/DSOfficeUtilities/Properties/AssemblyInfo.cs
@@ -1,16 +1,12 @@
 using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("DSOffice")]
+[assembly: AssemblyTitle("DSOfficeUtilities")]
 [assembly: AssemblyCulture("")]
 
+[assembly: InternalsVisibleTo("DSOffice")]
+[assembly: InternalsVisibleTo("DynamoCoreTests")]
 [assembly: InternalsVisibleTo("DynamoMSOfficeTests")]
-
-#if _WINDOWS
-    [assembly: TypeForwardedTo(typeof(DSOffice.WorkBook))]
-    [assembly: TypeForwardedTo(typeof(DSOffice.WorkSheet))]
-#endif
