+++ test/DynamoCoreTests/UnitTestBase.cs
@@ -65,10 +65,6 @@ namespace Dynamo
         public virtual void Setup()
         {
             SetupDirectories();
-#if NETFRAMEWORK
-            
-            DSOffice.ExcelInterop.ShowOnStartup = false;
-#endif
 
             if (assemblyHelper == null)
             {
