+++ test/Engine/FFITarget/ClassFunctionality.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Runtime.InteropServices;
 
 namespace FFITarget
 {
@@ -178,5 +179,14 @@ namespace FFITarget
         {
             throw new NotImplementedException();
         }
+    }
+
+    public class MissingClass
+    {
+        public static long MissingMethod([In, Optional] object value)
+        {
+            return (long)value;
+        }
+
     }
 }
